+++ src/frontend-cdoc-commons/components/cdoc-odobjectdetails/cdoc-odobjectdetails.component.spec.ts
@@ -1,0 +1,33 @@
+/* tslint:disable:no-unused-variable */
+import {async, ComponentFixture, TestBed} from '@angular/core/testing';
+import {NO_ERRORS_SCHEMA} from '@angular/core';
+import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
+import {TranslateModule} from '@ngx-translate/core';
+import {CommonDocODObjectDetailsComponent} from './cdoc-odobjectdetails.component';
+
+describe('TourDocODObjectDetailsComponent', () => {
+    let component: CommonDocODObjectDetailsComponent;
+    let fixture: ComponentFixture<CommonDocODObjectDetailsComponent>;
+
+    beforeEach(async(() => {
+        TestBed.configureTestingModule({
+            declarations: [CommonDocODObjectDetailsComponent],
+            schemas: [NO_ERRORS_SCHEMA],
+            imports: [NgbModule.forRoot(),
+                TranslateModule.forRoot()],
+            providers: [
+            ]
+        })
+            .compileComponents();
+    }));
+
+    beforeEach(() => {
+        fixture = TestBed.createComponent(CommonDocODObjectDetailsComponent);
+        component = fixture.componentInstance;
+        fixture.detectChanges();
+    });
+
+    it('should create', () => {
+        expect(component).toBeTruthy();
+    });
+});