+++ docs/CHANGELOG.md
@@ -19,6 +19,25 @@
 - frontend: export "real" angular-modules
 
 
+## 4.2.0
+- added new components for managing object-detection image-objects
+
+### new features
+- frontend: added new components for managing object-detection image-objects
+
+### improvements
+- none
+
+### bug fixes
+- none
+ 
+### breaking changes
+- none
+
+### known issues
+- frontend: IE till 11 not functional (can block rendering sometimes)
+
+
 ## 4.1.0
 - improved build-process
 - frontend: several improvements for maps, filters, itemlists